wip: mentions example get rid of warnings
