+++ src/dotnet/Chat.UI.Blazor/Components/SlateEditor/components.ts
@@ -1,0 +1,7 @@
+import ReactDOM from 'react-dom'
+
+export const Portal = ({ children }) => {
+    return typeof document === 'object'
+           ? ReactDOM.createPortal(children, document.body)
+           : null
+}
