+++ src/dotnet/Chat.UI.Blazor/Components/SlateEditor/slate-component.tsx
@@ -1,12 +1,12 @@
 // Import React dependencies.
 import React, { useState } from 'react'
 // Import the Slate editor factory.
-import { createEditor } from 'slate'
+import { createEditor, Descendant } from 'slate'
 
 // Import the Slate components and React plugin.
 import { Slate, Editable, withReact } from 'slate-react'
 
-const initialValue = [
+const initialValue: Descendant[] = [
     {
         type: 'paragraph',
         children: [{ text: 'A line of text in a paragraph.' }],
